Swet Alert 2 para servicio de usuario y login

+++ client/admininfo/src/services/usuarios/usuariostable.js
@@ -7,22 +7,85 @@ import { BsFillPenFill, BsFillTrashFill } from "react-icons/bs";
 import { deleteUsuario } from "../../routes/usuarios";
 import { AuthHeaders } from "../../components/authheader";
 
+import Swal from "sweetalert2";
+
 const TableUsuarios = (props) => {
-  const { noReg, user, rol, estado} = props.obj;
+  const { noReg, user, rol, estado } = props.obj;
 
   const navigate = useNavigate();
 
   const borrarUsuario = async () => {
-  
+    Swal.fire({
+      title: "Desea eliminar el usuario? ",
+      html: "Deleting will be canceled in 10 <strong></strong> seconds.",
+      timer: 10000,
+      timerProgressBar: true,
+      showDenyButton: true,
+      showConfirmButton: true,
+      confirmButtonText: "Delete",
+      denyButtonText: "Not delete",
+    }).then(async (result) => {
+      try {
+        if (result.isConfirmed) {
+          const authheader = AuthHeaders();
+          await deleteUsuario(user, authheader);
 
-    try {
-      const authheader = AuthHeaders();
-      await deleteUsuario(user, authheader);
-      console.log("Usuario Eliminado");
-      window.location.reload();
-    } catch (error) {
-      console.log("Usuario no se ha podido eliminar", error);
-    }
+          Swal.fire({
+            icon: "success",
+            title: "Usuario Eliminado",
+            showConfirmButton: false,
+            timer: 2000,
+            didOpen: () => {
+              Swal.showLoading();
+            },
+          });
+          setTimeout(() => {
+            Swal.close();
+            window.location.reload();
+          }, 2000);
+        } else {
+          if (result.isDenied) {
+            Swal.fire({
+              icon: "info",
+              title: "Usuario no ha sido Eliminado",
+              showConfirmButton: false,
+              timer: 2000,
+              didOpen: () => {
+                Swal.showLoading();
+              },
+            });
+            setTimeout(() => {
+              Swal.close();
+            }, 2000);
+          }
+        }
+        if (result.dismiss === Swal.DismissReason.timer) {
+          Swal.fire({
+            icon: "error",
+            title: "Se ha superado el tiempo sin una respuesta",
+            showConfirmButton: false,
+            timer: 2000,
+            didOpen: () => {
+              Swal.showLoading();
+            },
+          });
+          setTimeout(() => {
+            Swal.close();
+          }, 2000);
+        }
+      } catch (error) {
+        let mensaje;
+        mensaje = error.response.data;
+
+        Swal.fire({
+          icon: "error",
+          title: mensaje,
+          showConfirmButton: false,
+          timer: 2000,
+        });
+        Swal.showLoading();
+      }
+    });
   };
 
   return (
